@@ -8,7 +8,7 @@ module.exports = {
   // function for our authenticated routes
   authMiddleware: function ({ req, res, next }) {
     // allows token to be sent via  req.query or headers
-    let token = req.body.token || req.query.token || req.headers.authorization;
+    let token =  req.query.token || req.headers.authorization;
 
     // ["Bearer", "<tokenvalue>"]
     if (req.headers.authorization) {
@@ -29,7 +29,7 @@ module.exports = {
     }
 
     // send to next endpoint
-    next();
+    return req
   },
   signToken: function ({ username, email, _id }) {
     const payload = { username, email, _id };
